+++ navground_core/include/navground/core/property.h
@@ -16,6 +16,35 @@
 
 namespace navground::core {
 
+template <typename T> inline std::string_view field_type_name();
+template <> inline std::string_view field_type_name<int>() { return "int"; }
+template <> inline std::string_view field_type_name<ng_float_t>() {
+  return "float";
+}
+template <> inline std::string_view field_type_name<bool>() { return "bool"; }
+template <> inline std::string_view field_type_name<std::string>() {
+  return "str";
+}
+template <> inline std::string_view field_type_name<Vector2>() {
+  return "vector";
+}
+template <> inline std::string_view field_type_name<std::vector<int>>() {
+  return "[int]";
+}
+template <> inline std::string_view field_type_name<std::vector<ng_float_t>>() {
+  return "[float]";
+}
+template <> inline std::string_view field_type_name<std::vector<bool>>() {
+  return "[bool]";
+}
+template <>
+inline std::string_view field_type_name<std::vector<std::string>>() {
+  return "[str]";
+}
+template <> inline std::string_view field_type_name<std::vector<Vector2>>() {
+  return "[vector]";
+}
+
 struct HasProperties;
 
 /**
@@ -59,26 +88,7 @@ struct Property {
     return std::visit(
         [](auto &&arg) {
           using T = std::decay_t<decltype(arg)>;
-          if constexpr (std::is_same_v<T, int>)
-            return "int";
-          if constexpr (std::is_same_v<T, ng_float_t>)
-            return "float";
-          if constexpr (std::is_same_v<T, bool>)
-            return "bool";
-          if constexpr (std::is_same_v<T, std::string>)
-            return "str";
-          if constexpr (std::is_same_v<T, Vector2>)
-            return "vector";
-          if constexpr (std::is_same_v<T, std::vector<int>>)
-            return "[int]";
-          if constexpr (std::is_same_v<T, std::vector<ng_float_t>>)
-            return "[float]";
-          if constexpr (std::is_same_v<T, std::vector<bool>>)
-            return "[bool]";
-          if constexpr (std::is_same_v<T, std::vector<std::string>>)
-            return "[str]";
-          if constexpr (std::is_same_v<T, std::vector<Vector2>>)
-            return "[vector]";
+          return field_type_name<T>();
         },
         field);
   }